fix(vector): stamp bbox_source on no-boundaries exit; dedup bbox test helpers
Round-1 review + SonarCloud follow-ups:
- processor: the "No page boundaries" early return in generate_highlights now
  stamps vector_sync.bbox_source on the envelope span, so every exit path is
  consistent (round-1 review nit).
- tests: fold the per-page-cache tests into test_pdf_highlighter_bbox.py and
  reuse its _make_pdf/_page_boundaries helpers instead of duplicating them,
  clearing the SonarCloud >3% new-code duplication gate.
